requirements.txt
git+https://github.com/MasoniteFramework/masonite#3.0 # use version without autoload
masonite-logging>=1.0,<2.0
masonite-validation>=3.0,<4.0
masonite-inertia==0.2.0
masonite-js-routes
pendulum==1.5.1
# deployment
gunicorn==20.0.4
psycopg2-binary==2.8.6
